// epam-atithi/webapp/src/app/models/invitation.model.ts
export class Invitation {
    epamLocation: string;
    estimatedDuration: string;
    inTime: string;
    invitationId: number;
    outTime: string;
    timeHrs?: number;
    timeMins?: number;
    meridian?: string;
    poc: string;
    purpose: string;
    status: string;
    visitCompleted: boolean;
    visitDateTime: string;
    visitorId: number;
    visitorLocation: string;
    visitorResponse: string;
  }